requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.5
django-embed-video==1.4.10
idna==3.10
pillow==11.1.0
pycparser==2.22
python-decouple==3.8
razorpay==1.4.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
toolset==2019.6.1
tzdata==2025.1
urllib3==2.3.0
